Updated MSVC project and got rid of some more float->int conversion warnings.

+++ src/components/action_button.cxx
@@ -103,14 +103,14 @@ VerticalActionButton::draw (DrawingContext& gc)
     {
       if (fast_mode) 
         {
-          gc.draw_fillrect(x_pos, y_pos, x_pos + 60, y_pos + 35,
+          gc.draw_fillrect((float)x_pos, (float)y_pos, (float)x_pos + 60, (float)y_pos + 35,
                            CL_Color(255, 255, 255));
         } 
       else 
         {
           //CL_Display::fill_rect(x_pos, y_pos, x_pos + 60, y_pos + 35 ,
           //1.0, 1.0, 1.0, 0.5);
-          gc.draw(backgroundhl, x_pos, y_pos);
+          gc.draw(backgroundhl, (float)x_pos, (float)y_pos);
         }
     }
   else
@@ -120,11 +120,11 @@ VerticalActionButton::draw (DrawingContext& gc)
       if (fast_mode) {
         // do nothing
       } else {
-        gc.draw(background, x_pos, y_pos);
+        gc.draw(background, (float)x_pos, (float)y_pos);
       }
     }
 
-  gc.draw(sprite, Vector(x_pos + 20, y_pos + 32));
+  gc.draw(sprite, Vector((float)x_pos + 20, (float)y_pos + 32));
 
   CL_Font myfont  = font;
   CL_Font myfontb = font_b;
@@ -136,19 +136,19 @@ VerticalActionButton::draw (DrawingContext& gc)
       && CL_Mouse::get_x() > x_pos      && CL_Mouse::get_x() < x_pos + 60
       && CL_Mouse::get_y() < y_pos + 35 && CL_Mouse::get_y() > y_pos)
     {
-      gc.print_left(myfontb, x_pos + 65, y_pos, action_to_screenname(name));
+      gc.print_left(myfontb, (float)x_pos + 65, (float)y_pos, action_to_screenname(name));
     }
 
 
   if (Cheat::unlimited_actions)
     {
       // FIXME: insert unlimited surface here
-      gc.print_center(myfont, x_pos + 46, y_pos + 5, "oo");
+      gc.print_center(myfont, (float)x_pos + 46, (float)y_pos + 5, "oo");
     }
   else
     {
       std::string str = CL_String::to(action_holder->get_available(name));
-      gc.print_center(myfont, x_pos + 46, y_pos + 5, str);
+      gc.print_center(myfont, (float)x_pos + 46, (float)y_pos + 5, str);
     }
 }
 
@@ -170,16 +170,16 @@ ArmageddonButton::draw (DrawingContext& gc)
 {
   if (server->get_world()->check_armageddon ())
     {
-      gc.draw(backgroundhl, Vector(x_pos, y_pos));
-      gc.draw(sprite, Vector(x_pos, y_pos));
+      gc.draw(backgroundhl, Vector((float)x_pos, (float)y_pos));
+      gc.draw(sprite, Vector((float)x_pos, (float)y_pos));
     }
   else
     {
       if (!fast_mode)
-        gc.draw(background, Vector(x_pos, y_pos));
+        gc.draw(background, Vector((float)x_pos, (float)y_pos));
 
       sprite.set_frame(7);
-      gc.draw(sprite, Vector(x_pos, y_pos));
+      gc.draw(sprite, Vector((float)x_pos, (float)y_pos));
     }
 }
 
@@ -248,15 +248,15 @@ ForwardButton::draw (DrawingContext& gc)
 {
   if (server->get_fast_forward())
     {
-      gc.draw(backgroundhl, Vector(x_pos, y_pos));
+      gc.draw(backgroundhl, Vector((float)x_pos, (float)y_pos));
     }
   else
     {
       if (!fast_mode)
-        gc.draw(background, Vector(x_pos, y_pos));
+        gc.draw(background, Vector((float)x_pos, (float)y_pos));
     }
 
-  gc.draw(surface, Vector(x_pos, y_pos));
+  gc.draw(surface, Vector((float)x_pos, (float)y_pos));
 }
 
 bool
@@ -296,15 +296,15 @@ PauseButton::draw (DrawingContext& gc)
 {
   if (server->get_pause())
     {
-      gc.draw(backgroundhl, Vector(x_pos, y_pos));
+      gc.draw(backgroundhl, Vector((float)x_pos, (float)y_pos));
     }
   else
     {
     if (!fast_mode)
-      gc.draw(background, Vector(x_pos, y_pos));
+      gc.draw(background, Vector((float)x_pos, (float)y_pos));
     }
 
-  gc.draw(surface, Vector(x_pos, y_pos));
+  gc.draw(surface, Vector((float)x_pos, (float)y_pos));
 }
 
 bool
